splitted the server.js with the new api endpoints
