+++ server.js
@@ -368,698 +368,6 @@ app.put('/server/settings', async (req, res) => {
   if (req.body.data.status === 'Modify an Author') { await updateAuthor(req, res); }
 })
 
-/*
-UPDATED API
-*/
-
-/**
- * @openapi
- * /api/authors:
- *  get:
- *    description: Get a list of authors, paginated. by default it's page 1 with size 1. Currently pages are broken
- *    responses:
- *      200:
- *        description: A list of authors
- */
-app.get('/api/authors', async (req, res) => {
-  const page = req.query.page;
-  const size = req.query.size;
-  if(page == undefined)
-    page = 1;
-  if(size == undefined)
-    size = 5;
-  const sanitizedAuthors = await getAuthors(page, size);
-
-  return res.json({
-    "type": "authors",
-    "items": [sanitizedAuthors]
-  });
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}:
- *  get:
- *    description: Fetchs a single Author object from the database and sends it back as a JSON object
- *    responses:
- *      404:
- *        description: Returns Status 404 when an Author does not exist 
- *      500:
- *        description: Returns Status 500 when the server is unable to retrieve the Author from the database
- */
-app.get('/api/authors/:authorId', async (req, res) => {
-  if(req.params.authorId == undefined)
-    return res.sendStatus(404);
-
-  let author = await getAuthor(req.params.authorId);
-
-  if(author === 404)
-    return res.sendStatus(404);
-
-  if(author === 500)
-    return res.sendStatus(500);
-
-  return res.json({
-    "type": "author",
-    "id" : author._id,
-    "host": process.env.DOMAIN_NAME,
-    "displayname": author.username,
-    "url":  process.env.DOMAIN_NAME + "users/" + author._id,
-    "github": "",
-    "profileImage": "",
-    "email": author.email, 
-    "about": author.about,
-    "pronouns": author.pronouns
-  });
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}:
- *  post:
- *    description: Gets the author's cookies from the database and sends the status
- *    responses:
- *      404:
- *        description: Returns Status 404 when the cookies don't exist
- *      400:
- *        description: Returns Status 400 when the body type doesn't match the author 
- *      400:
- *        description: Returns Status 400 when the author ID, host, nor username are valid 
- */
-app.post('/api/authors/:authorId', async (req, res) => {
-  if(!req.cookies["token"])
-    return res.sendStatus(401);
-  if(req.body.type !== 'author')
-    return res.sendStatus(400);
-
-  const authorId = req.body.id;
-  const host = req.body.host;
-  const username = req.body.displayName;
-
-  if(!authorId || !host || !username)
-    return res.sendStatus(400);
-
-  return res.sendStatus(await apiUpdateAuthor(req.cookies["token"], req.body));
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/followers:
- *  get:
- *    description: Gets the author's followers and friends as objects from the database and sends it back as a JSON object
- *    responses:
- *      404:
- *        description: Returns Status 404 when there are no existing followers or friends
- */
-app.get('/api/authors/:authorId/followers', async (req, res) => {
-  const authorId = req.params.authorId;
-
-  const followers = await getFollowers(authorId);
-  const friends = await getFriends(authorId);
-
-  if(followers == 404 || friends == 404)
-    return res.send(404);
-
-  santizedFollowers = [];
-  for(let i = 0; i < followers.length; i++){
-    const follower = followers[i];
-
-    const followerProfile = await Author.findOne({_id: follower.authorId}); 
-    if(!followerProfile)
-      continue
-
-    santizedFollower = {
-      "type": "author",
-      "id" : followerProfile._id,
-      "host": process.env.DOMAIN_NAME,
-      "displayname": followerProfile.username,
-      "url":  process.env.DOMAIN_NAME + "users/" + followerProfile._id,
-      "github": "",
-      "profileImage": "",
-      "email": followerProfile.email,
-      "about": followerProfile.about,
-      "pronouns": followerProfile.pronouns
-    }
-
-    santizedFollowers.push(santizedFollower);
-  }
-
-  for(let i = 0; i < friends.length; i++){
-    const friend = friends[i];
-
-    const friendProfile = await Author.findOne({_id: friend.authorId}); 
-    if(!friendProfile)
-      continue
-    santizedFollower = {
-      "type": "author",
-      "id" : friendProfile._id,
-      "host": process.env.DOMAIN_NAME,
-      "displayname": friendProfile.username,
-      "url":  process.env.DOMAIN_NAME + "users/" + friendProfile._id,
-      "github": "",
-      "profileImage": "",
-      "about": friendProfile.about,
-      "pronouns": friendProfile.pronouns
-    }
-
-    santizedFollowers.push(santizedFollower);
-  }
-
-  return res.json({
-    type: "followers",
-    items: santizedFollowers
-  });
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/followers/{foreignAuthorId}:
- *  get:
- *    description: Gets the author's followers' and friends' profiles as objects from the database and sends it back as a JSON object
- *    responses:
- *      404:
- *        description: Returns Status 404 when the follower or friend object is not found
- *      404:
- *        description: Returns Status 404 when the profile doesn't exist
- */
-app.get('/api/authors/:authorId/followers/:foreignAuthorId', async (req, res) => {
-  const authorId = req.params.authorId;
-  const foreignId = req.params.foreignAuthorId;
-
-  const followers = await getFollowers(authorId);
-  const friends = await getFriends(authorId);
-
-  if(followers == 404 || friends == 404)
-    return res.send(404);
-
-  for(let i = 0; i < followers.length; i++){
-    const follower = followers[i];
-    if(follower.authorId == foreignId){
-
-      const followerProfile = await Author.findOne({_id: follower.authorId}); 
-      if(!followerProfile)
-        continue
-
-      return res.json({
-        "type": "author",
-        "id" : followerProfile._id,
-        "host": process.env.DOMAIN_NAME,
-        "displayname": followerProfile.username,
-        "url":  process.env.DOMAIN_NAME + "users/" + followerProfile._id,
-        "github": "",
-        "profileImage": "",
-        "about": followerProfile.about,
-        "pronouns": followerProfile.pronouns
-      });
-    }
-  }
-
-  for(let i = 0; i < friends.length; i++){
-    const friend = friends[i];
-    if(friend.authorId = foreignId){
-
-      const friendProfile = await Author.findOne({_id: friend.authorId}); 
-
-      if(!friendProfile)
-        continue
-
-      return res.json({
-        "type": "author",
-        "id" : friendProfile._id,
-        "host": process.env.DOMAIN_NAME,
-        "displayname": friendProfile.username,
-        "url":  process.env.DOMAIN_NAME + "users/" + friendProfile._id,
-        "github": "",
-        "profileImage": "",
-        "about": friendProfile.about,
-        "pronouns": friendProfile.pronouns
-      })
-    }
-  }
-
-  return res.sendStatus(404);
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/followers/{foreignAuthorId}:
- *  put:
- *    description: Adds a Follower (Foreign Author) to the Author's Follower List in the database
- *    responses:
- *      401:
- *        description: Returns Status 401 when the adding the Follower is not authorized 
- *      400:
- *        description: Returns Status 400 when server is unable to process the user's request 
- */
-app.put('/api/authors/:authorId/followers/:foreignAuthorId', async (req, res) => {
-
-  const authorId = req.params.authorId;
-  const foreignId = req.params.foreignAuthorId;
-
-  const follower = await addFollower(req.cookies["token"], authorId, foreignId, req.body, req, res);
-
-  if(follower == 401)
-    return res.sendStatus(401);
-  else if(follower == 400)
-    return res.sendStatus(400);
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/followers/{foreignAuthorId}:
- *  delete:
- *    description: Deletes a Follower (Foreign Author) to Author 
- *    responses:
- *      400:
- *        description: If the request has no type or if the type is not a follow request
- *      200:
- *        description: If the Follower (Foreign Author) was successfully deleted 
- */
-app.delete('/api/authors/:authorId/followers/:foreignAuthorId', async (req, res) => {
-  if(req.body.type == undefined || req.body.type != "follower")
-    return res.sendStatus(400)
-
-  const authorId = req.params.authorId;
-  const foreignId = req.params.foreignAuthorId;
-
-  const statusCode = await deleteFollower(req.cookies["token"], authorId, foreignId, req.body);
-
-  return res.sendStatus(statusCode);
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/requests/{foreignAuthorId}:
- *  put:
- *    description: Saves the Request for the Foreign Author from the Author into the database 
- *    responses:
- *      200:
- *        description: Returns the JSON object representing the Request  
- */
-app.put('/api/authors/:authorId/requests/:foreignAuthorId', async (req, res) => {
-  const authorId = req.params.authorId;
-  const foreignId = req.params.foreignAuthorId;
-
-  const request = await sendRequest(authorId, foreignId, res);
-
-  return res.json({
-    "type": request.type,
-    "summary": request.summary,
-    "actor": request.actor,
-    "object": request.object
-  })
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/requests/{foreignAuthorId}:
- *  delete:
- *    description: Deletes a Request made by the Author  to Foreign Author
- *    responses:
- *     200:
- *        description: Returns the JSON object representing the Request 
- */
-app.delete('/api/authors/:authorId/requests/:foreignAuthorId', async (req, res) => {
-  const authorId = req.params.authorId;
-  const foreignId = req.params.foreignAuthorId;
-
-  const request = await apideleteRequest(authorId, foreignId, res);
-
-  return res.json({
-    "type": request.type,
-    "summary": request.summary,
-    "actor": request.actor,
-    "object": request.object
-  })
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts/{postId}:
- *  get:
- *    description: Gets the a specific Post from the database made by a specific Author 
- *    responses:
- *      200:
- *        description: Returns the Post as a JSON object
- *      404:
- *        description: When the Post or Author is not found 
- *      500: 
- *        description: When the server is unable to get the Post
- */
-app.get('/api/authors/:authorId/posts/:postId', async (req, res) => {
-  if(req.params.authorId == undefined) return res.sendStatus(404);
-  const authorId = req.params.authorId;
-  const postId = req.params.postId;
-
-  let post = await apigetPost(authorId, postId);
-
-  if(post === 404) return res.sendStatus(404);
-  if(post === 500) return res.sendStatus(500);
-
-  return res.json({
-    "type": "post",
-    "title" : post.title,
-    "id": process.env.DOMAIN_NAME + "authors/" + authorId + "/" + postId,
-    "source": post.source,
-    "origin": post.origin,
-    "description": post.description,
-    "contentType": post.contentType,
-    "author": post.author, 
-    "categories": post.categories,
-    "count": post.count,
-    "comments": post.comments,
-    "commentSrc": post.commentSrc,
-    "published": post.published,
-    "visibility": post.visibility,
-    "unlisted": post.unlisted
-  });
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts/{postId}:
- *  post:
- *    description: Updates an existing Post made by a specific Author 
- *    responses:
- *      200:
- *        description: Returns Status 200 when the server successfully updated the Post
- *      404:
- *        description: Returns Status 404 when the server was unable to find any Post related to the given URL
- */
-app.post('/api/authors/:authorId/posts/:postId', async (req, res) => {
-  const authorId = req.params.authorId;
-  const postId = req.params.postId;
-
-  const status = await apiupdatePost(authorId, postId, req.body);
-  
-  if (status == 200) {
-    return res.sendStatus(status);
-  } else {
-    return res.sendStatus(404);
-  }
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts/{postId}:
- *  delete:
- *    description: Deletes a specific Post made by a specific Author
- *    responses:
- *      200:
- *        description: Returns Status 200 if the server successfully deleted the Post
- *      404:
- *        description: Returns Status 404 if the server was unable to find either the Author or the Author's Post
- *      500:
- *        description: Returns Status 500 if the server was unable to delete the Post
- */
-app.delete('/api/authors/:authorId/posts/:postId', async (req, res) => {
-  const authorId = req.params.authorId;
-  const postId = req.params.postId;
-
-  const status = await apideletePost(authorId, postId);
-
-  if (status == 200) {
-    return res.sendStatus(status);
-  } else if (status == 404) {
-    return res.sendStatus(404);
-  } else if (status == 500) {
-    return res.sendStatus(500);
-  }
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts/{postId}:
- *  put:
- *    description: Creates a Post object made by a specific Author 
- *    responses:
- *      200:
- *        description: Returns Status 200 if the Post was able to be stored into the database 
- *      404:
- *        description: Returns Status 404 if the Author or Post does not exist 
- *      500: 
- *        description: Returns Status 500 if the server was unable to create and save the Post
- */
-app.put('/api/authors/:authorId/posts/:postId', async (req, res) => {
-  const authorId = req.params.authorId;
-  const postId = req.params.postId;
-
-  const status = await apicreatePost(authorId, postId, req.body, process.env.DOMAIN_NAME);
-
-  if (status == 200) {
-    return res.sendStatus(status);
-  } else if (status == 404) {
-    return res.sendStatus(404);
-  } else if (status == 500) {
-    return res.sendStatus(500); 
-  }  
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts:
- *  get:
- *    description: Fetches the posts of a specific Author 
- *    responses:
- *      200:
- *        description: Returns a list of Posts by a specific Author
- */
-app.get('/api/authors/:authorId/posts', async (req, res) => {
-  const authorId = req.params.authorId;
-  const page = req.query.page;
-  const size = req.query.size;
-
-  if(page == undefined)
-  page = 1;
-  if(size == undefined)
-    size = 5;
-
-  const sanitizedPosts = await fetchPosts(page, size, authorId);
-
-  return res.json({
-    "type": "posts",
-    "authorId": authorId,
-    "items": [sanitizedPosts]
-  });
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts:
- *  post:
- *    description: Creates a Post given no PostId (i.e., needs to be generated)
- *    responses:
- *      200:
- *        description: Returns Status 200 if the Post was successfully created and saved into the database
- *      404:
- *        description: Returns Status 404 if the Author does not exist 
- *      500 
- *        description: Returns Status 500 if the server was unable to store or create the Post 
- */
-app.post('/api/authors/:authorId/posts', async (req, res) => {
-  const authorId = req.params.authorId;
-
-  const status = await apicreatePost(authorId, undefined, req.body, process.env.DOMAIN_NAME);
-
-  if (status == 200) {
-    return res.sendStatus(status);
-  } else if (status == 404) {
-    return res.sendStatus(404);
-  } else if (status == 500) {
-    return res.sendStatus(500); 
-  }  
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts/{postId}/comments:
- *  get:
- *    description: Gets the comments related to a specific Post made by a specific Author (paginated)
- *    responses:
- *      200:
- *        description: Returns Status 200 when the comments have been successfully found (i.e., returns list of comments)
- */
-app.get('/api/authors/:authorId/posts/:postId/comments', async (req, res) => {
-  const authorId = req.params.authorId;
-  const postId = req.params.postId;
-  let page = req.query.page;
-  let size = req.query.size;
-
-  if(page == undefined)
-  page = 1;
-  if(size == undefined)
-    size = 5;
-
-  const comments = getComments(authorId, postId);
-
-  return res.json({
-    "type": "comments",
-    "page": page,
-    "size": size,
-    "post": process.env.DOMAIN_NAME + "/authors/" + authorId + "/posts/" + postId,
-    "id": process.env.DOMAIN_NAME + "/authors/" + authorId + "/posts/" + postId + "/comments",
-    "comments": comments
-    })
-})
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts/{postId}/comments:
- *  post:
- *    description: Creates a comment for a specific Post made by a specific Author
- *    responses:
- *      200:
- *        description: Returns 200 when the comment was successfully made (i.e., returns the comment)
- */
-app.post('/api/authors/:authorId/posts/:postId/comments', async (req, res) => {
-  const authorId = req.params.authorId;
-  const postId = req.params.postId;
-
-  const comment = createComment(authorId, postId, req.body, process.env.DOMAIN_NAME);
-
-  return res.json({
-    "type": "comment",
-    "author": comment.author,
-    "comment": comment.comment,
-    "contentType": comment.contentType,
-    "published": comment.published,
-    "id": comment._id
-    }) 
-})
-
-
-/**
- * @openapi
- * /api/authors/{authorId}/posts/{postId}/likes:
- *  get:
- *    description: Fetches the likes related to a specific Post made by a specific Author (paginated)
- *    responses:
- *      200:
- *        description: Returns 200 if the likes were successfully fetched (i.e., return the likes )
- */
-app.get('/api/authors/:authorId/posts/:postId/likes', async (req, res) => {
-  const authorId = req.params.authorId;
-  const postId = req.params.postId;
-  let page = req.query.page;
-  let size = req.query.size;
-
-  if(page == undefined)
-  page = 1;
-  if(size == undefined)
-    size = 5;
-  
-  const likes = apifetchLikes(authorId, postId);
-
-  return res.json({
-    "type": "likes",
-    "page": page, 
-    "size": size,
-    "post": process.env.DOMAIN_NAME + "/authors/" + authorId + "/posts/" + postId,
-    "id": process.env.DOMAIN_NAME + "/authors/" + authorId + "/posts/" + postId + "/likes",
-    "likes": likes
-    })
-
-})
-
-/**
- * @openapi
- * /api/authors/:authorId/posts/:postId/comments/:commentId/likes:
- *  get:
- *    description: Fetches the likes on a specific comment on a specific post by a specific author
- *    responses:
- *      200:
- *        description: Returns 200 if server retrieves the likes for the comment on a post 
- */
-app.get('/api/authors/:authorId/posts/:postId/comments/:commentId/likes', async (req, res) => {
-  const authorId = req.params.authorId;
-  const postId = req.params.postId;
-  const commentId = req.params.commentId
-  let page = req.query.page;
-  let size = req.query.size;
-
-  if(page == undefined)
-  page = 1;
-  if(size == undefined)
-    size = 5;
-
-  const likes = apiFetchCommentLikes(authorId, postId, commentId); 
-
-  return res.json({
-    "type": "likes",
-    "page": page,
-    "size": size,
-    "post": process.env.DOMAIN_NAME + "/authors/" + authorId + "/posts/" + postId,
-    "comment": process.env.DOMAIN_NAME + "/authors/" + authorId + "/posts/" + postId + "/comments/" + commentId,
-    "id": process.env.DOMAIN_NAME + "/authors/" + authorId + "/posts/" + postId + "/comments" + commentId + "/likes",
-    "likes": likes
-    })
- })
-
-/**
- * @openapi
- * /api/authors/:authorId/liked:
- *  get:
- *    description: Fetches all the likes the author has made 
- *    responses:
- *      200:
- *        description: Returns 200 with the list of likes the author has made
- */
-app.get('/api/authors/:authorId/liked', async (req, res) => {
-  const authorId = req.params.authorId;
-
-  const liked = getAuthorLikes(authorId);
-
-  return res.json({
-      "type":"liked",
-      "items": liked
-  })
-})
-
-/**
- * @openapi
- * /server/authors/:author_id/inbox:
- *  get:
- *    description: Gets the Author's Inbox
- *    responses:
- *      200:
- *        description: If the server finds an Inbox for the specified Author
- */
-app.get('/server/authors/:author_id/inbox', async (req, res) => {
-  console.log('Debug: Getting an authors inbox');
-  await getInbox(req, res);
-})
-
-/**
- * @openapi
- * /server/authors/:author_id/inbox:
- *  get:
- *    description: Adds a Post, Follow, Comment, or Like to the Inbox of an Author
- *    responses:
- *      200:
- *        description: If the server successfully adds either a Post, Follow, Comment, or Like to the Author's Inbox
- */
-app.post('/server/authors/:author_id/inbox', async (req, res) => {
-  await postInbox(req, res);
-})
-
-/**
- * @openapi
- * /server/authors/:author_id/inbox:
- *  get:
- *    description: Deletes the Inbox
- *    responses:
- *      200:
- *        description: If the server finds an Inbox for the specified Author
- */
-app.delete('/server/authors/:author_id/inbox', async (req, res) => {
-  console.log("delete inbox")
-  await deleteInbox(req, res);
-})
-
-/*
-END OF NEW API STUFF
-*/
-
 app.get('/',(req, res) => {
   res.render("index");
 });
